soowa/requirements.txt
absl-py==0.13.0
asgiref==3.3.4
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
deepdiff==5.5.0
Django==3.2.3
djangorestframework==3.12.4
enum34==1.1.10
future==0.18.2
idna==2.10
mediapipe==0.8.5
mock==4.0.3
mpyq==0.2.5
numpy==1.20.3
opencv-contrib-python==4.5.2.54
opencv-python==4.5.2.54
ordered-set==4.0.2
Pillow==8.2.0
portpicker==1.4.0
protobuf==3.17.3
pygame==2.0.1
PySC2==3.0.0
pytz==2021.1
requests==2.25.1
runtime==0.1.4
s2clientprotocol==5.0.7.84643.0
s2protocol==5.0.7.84643.0
scipy==1.6.3
six==1.16.0
sk-video==1.1.10
sqlparse==0.4.1
urllib3==1.26.5
websocket-client==1.1.0
whichcraft==0.6.1
absl-py==0.13.0
asgiref==3.3.4
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
deepdiff==5.5.0
Django==3.2.3
djangorestframework==3.12.4
enum34==1.1.10
future==0.18.2
idna==2.10
mediapipe==0.8.5
mock==4.0.3
mpyq==0.2.5
numpy==1.20.3
opencv-contrib-python==4.5.2.54
opencv-python==4.5.2.54
ordered-set==4.0.2
Pillow==8.2.0
portpicker==1.4.0
protobuf==3.17.3
pygame==2.0.1
PySC2==3.0.0
pytz==2021.1
requests==2.25.1
runtime==0.1.4
s2clientprotocol==5.0.7.84643.0
s2protocol==5.0.7.84643.0
scipy==1.6.3
six==1.16.0
sk-video==1.1.10
sqlparse==0.4.1
urllib3==1.26.5
uWSGI==2.0.19.1
websocket-client==1.1.0
whichcraft==0.6.1
absl-py==0.13.0
asgiref==3.3.4
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
deepdiff==5.5.0
Django==3.2.3
djangorestframework==3.12.4
enum34==1.1.10
future==0.18.2
idna==2.10
mediapipe==0.8.5
mock==4.0.3
mpyq==0.2.5
numpy==1.20.3
opencv-contrib-python==4.5.2.54
opencv-python==4.5.2.54
ordered-set==4.0.2
Pillow==8.2.0
portpicker==1.4.0
protobuf==3.17.3
pygame==2.0.1
PySC2==3.0.0
pytz==2021.1
requests==2.25.1
runtime==0.1.4
s2clientprotocol==5.0.7.84643.0
s2protocol==5.0.7.84643.0
scipy==1.6.3
six==1.16.0
sk-video==1.1.10
sqlparse==0.4.1
urllib3==1.26.5
uWSGI==2.0.19.1
websocket-client==1.1.0
whichcraft==0.6.1
absl-py==0.13.0
asgiref==3.3.4
astunparse==1.6.3
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
deepdiff==5.5.0
Django==3.2.3
djangorestframework==3.12.4
enum34==1.1.10
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==1.32.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==2.10
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
Markdown==3.3.4
mediapipe==0.8.5
mock==4.0.3
mpyq==0.2.5
numpy==1.19.5
oauthlib==3.1.1
opencv-contrib-python==4.5.2.54
opencv-python==4.5.2.54
opt-einsum==3.3.0
ordered-set==4.0.2
Pillow==8.2.0
portpicker==1.4.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.0.1
PySC2==3.0.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
runtime==0.1.4
s2clientprotocol==5.0.7.84643.0
s2protocol==5.0.7.84643.0
scipy==1.6.3
six==1.15.0
sk-video==1.1.10
sqlparse==0.4.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.5
uWSGI==2.0.19.1
websocket-client==1.1.0
Werkzeug==2.0.1
whichcraft==0.6.1
wrapt==1.12.1
